requirements.txt
cohere
pandas
python-dotenv
colorama